Appease Clippy-sama, remove stuff
